Generalize Rusanov to be problem-indipendent
We provide a function `eigs` associated to the specific problem that
returns the spectral space (i.e. the proper eigenvalues) of the problem
we are solving.

Theoretically we could provide whatever eigs in order to provide
different spectral behaviours to be used by schemes.

Probably instead of `eigs`, we could also provide the matrix A