Enable pickling of entire Solver class
This allows to deepcopy the Solver class directly. That means now in the
MemoryWriter we can just store copy of the Solver instances such that
it's more natural to use the stored data instead of learning another
data structure like `josie.data.StateElement`.

I plan to directly copy the solver object also in the plot module such
that we don't need all the shenanigans to convert from Solver ->
StateElement and viceversa